Initial Commit
This is the initial commit, where I am able to create the visibility graph and calculate the hurst exponent for a small dataset of Google stocks. But the visibility graph and the weights are not provided correctly. So, we have very faulty results. This needs to be solved in the future.